backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-chess==1.999
python-multipart==0.0.20
openai==1.58.1
pydantic==2.10.4
pydantic-settings==2.7.1
httpx==0.28.1
